# z/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Opinion Analytics API",
    "version": "1.0.0",
    "description": "AI 驱动的视频评论分析平台后端接口，支持 B站/抖音 评论采集、Embedding 向量化、聚类分析、Topic 生成、AI 总结等。"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "本地开发" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "description": "0=成功, 401=未登录, 403=权限不足, 404=资源不存在, 429=限流, 500=服务端错误" },
          "message": { "type": "string" }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "user_id": { "type": "integer" },
          "username": { "type": "string" },
          "display_name": { "type": "string", "description": "显示名称" },
          "role": { "type": "string", "description": "角色" },
          "avatar_initial": { "type": "string", "description": "头像首字母" },
          "email": { "type": "string", "format": "email" }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" },
          "remember": { "type": "boolean", "default": false }
        }
      },
      "LoginData": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserProfile" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 3 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "example": {
                "username": "admin",
                "password": "password",
                "remember": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/LoginData" } } }
                  ]
                },
                "example": {
                  "code": 0,
                  "data": {
                    "token": "eyJhbGciOiJIUzI1NiIs...",
                    "user": { "user_id": 1, "username": "admin", "display_name": "刘心宇", "role": "管理员", "avatar_initial": "L" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "example": { "code": 401, "message": "用户名或密码错误" }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["认证"],
        "summary": "用户注册",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "example": {
                "username": "newuser",
                "email": "user@example.com",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "example": { "code": 0, "message": "注册成功" }
              }
            }
          }
        }
      }
    },
...
  }
}